[CI/Build] [2/3] Reorganize entrypoints tests (#5904)
