+++ tests/entrypoints/test_openai_chat.py
@@ -72,7 +72,7 @@ TEST_CHOICE = [
 pytestmark = pytest.mark.openai
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="module")
 def zephyr_lora_files():
     return snapshot_download(repo_id=LORA_NAME)
